==> notebooks/01-jrsotc-build-val-dataset.ipynb <==
import pandas as pd 
import numpy as np
from sklearn.preprocessing import LabelEncoder

### validate
df_val = pd.read_csv("../input/jigsaw-toxic-severity-rating/validation_data.csv")
print (f"df_val:{df_val.shape[0]}")
texts =np.concatenate ( [df_val["less_toxic"].values,df_val["more_toxic"].values] )

df_text = pd.DataFrame ({
    "text":list(set(texts))
})

print (f"df_text:{df_text.shape[0]}")
df_text.to_csv("../processed/validation_texts.csv", index=False)

df_text.head()

text_encoder = LabelEncoder ()
text_encoder.fit(texts)

df_val["less_toxic"]=text_encoder.transform(df_val["less_toxic"])
df_val["more_toxic"]=text_encoder.transform(df_val["more_toxic"])
df_val["text_a"] = df_val.apply(lambda row: row["less_toxic"] if row["less_toxic"]  <  row["more_toxic"] else  row["more_toxic"],axis=1).astype(int)
df_val["text_b"] = df_val.apply(lambda row: row["more_toxic"] if row["less_toxic"]  <  row["more_toxic"] else  row["less_toxic"],axis=1).astype(int)
df_val["win_a"] = df_val.apply(lambda row: 0 if row["less_toxic"]  <  row["more_toxic"] else 1,axis=1)
df_val["win_b"] = df_val.apply(lambda row: 1 if row["less_toxic"]  <  row["more_toxic"] else 0,axis=1)
df_val_agg = df_val.groupby(["text_a","text_b"]).agg({"win_a":"sum","win_b":"sum"}).reset_index()

df_val_agg["less_toxic"] = df_val_agg.apply (lambda row: row["text_a"] if row["win_a"]< row["win_b"] else row["text_b"], axis=1 )
df_val_agg["more_toxic"] = df_val_agg.apply (lambda row: row["text_b"] if row["win_a"]< row["win_b"] else row["text_a"], axis=1 )
df_val_agg["count"] = df_val_agg["win_a"]+df_val_agg["win_b"]
df_val_agg["count"] = df_val_agg["win_a"]+df_val_agg["win_b"]
df_val_agg["perc_win_a"] = df_val_agg["win_a"]/df_val_agg["count"]
df_val_agg["perc_win_b"] = df_val_agg["win_b"]/df_val_agg["count"]
df_val_agg["avg_agreement"] = df_val_agg[["perc_win_a","perc_win_b"]].max(axis=1)
df_val_agg["less_toxic"] = text_encoder.inverse_transform ( df_val_agg["less_toxic"] )
df_val_agg["more_toxic"] = text_encoder.inverse_transform ( df_val_agg["more_toxic"] )
df_val_agg = df_val_agg [["less_toxic", "more_toxic", "count", "avg_agreement"]]


df_val_agg.to_csv("../processed/validation.csv", index=False)

df_val_agg.head()
