==> notebooks/04b-jrsotc-feature-extraction-tf-idf-jc.ipynb <==
import pandas as pd
pd.set_option('display.max_colwidth', 512)

import numpy as np
import matplotlib.pyplot as plt
import pickle
from tqdm import tqdm
from sklearn.preprocessing import LabelEncoder
import cudf 
import cuml 


### Embedding

embeddings = pickle.load( open( "../processed/jc_embeddings.pickle", "rb" ) )

jc = pd.read_csv("../processed/jc.csv")


## KNN

%%time 

KNN = 500
model = cuml.neighbors.NearestNeighbors(n_neighbors=KNN)
model.fit(embeddings)
distances, indices = model.kneighbors(embeddings)

with open('../processed/jc_knn_distances.pickle', 'wb') as handle:
    pickle.dump(distances, handle)

with open('../processed/jc_knn_indices.pickle', 'wb') as handle:
    pickle.dump(indices, handle)

distances.shape, indices.shape

text = jc["text"].values
toxic = jc["toxic"].values
jc.shape

CUTOFF = 0.15
mm = distances[:,1]
idx = np.where( (mm<CUTOFF) )[0]

group_id = 0

d = {}
inv_d = {}
for i in idx:
    
    if i in inv_d:
        mark = inv_d[i]
    else:
        mark = i
        d[mark] = []
    
    for j in range(distances.shape[1]):
        if distances [i,j] < CUTOFF:
            k = indices[i,j]
            if k not in d[mark]: 
                d[mark].append ( k)
            inv_d [k] = mark
        else:
            break
group_ids = {} 
for n, i in enumerate(d):
    
    for j in d[i]:
        group_ids[j] = n

        
def find_intersection (d):
    count = 0
    for i in d:
        for k in d:
            if i <  k:
                s1=set(d[i])
                s2=set(d[k])
                if len(s1.intersection(s2))>0:
                    count += 1
    return count
        
def aggregate (d):
    d2 = {}
    founds = []
    for i in d:
        if i not in founds:
            d2[i] = d[i]
            founds.append(i)
            for k in d:
                if i <  k:
                    s1=set(d[i])
                    s2=set(d[k])
                    if len(s1.intersection(s2))>0:
                        d2[i] = d[i] + d[k]
                        founds.append(k)    
    return d2


count = find_intersection(d)
while count >0:
    print (count)
    d = aggregate (d)
    count = find_intersection(d)
    

group_ids = {}
for n, i in enumerate(d):
    l = len(d[i])
    for j in d[i]:
        group_ids[j] = -n


label_encoder = LabelEncoder()

jc["group_id"] = jc.index.map(lambda x: group_ids[x] if x in group_ids else np.nan )
jc["group_id"] = jc["group_id"].fillna(jc.index.to_series())
jc["group_id"] = label_encoder.fit_transform(jc["group_id"])


df_g = jc.groupby(["group_id"]).agg ({
    "id" :"count",
    "toxic":"std", 	"severe_toxic":"std", 	"obscene":"std", 	"threat":"std", 	"insult":"std", 	"identity_hate":"std",
}
).reset_index().rename(columns={"id":"total"})

df_g["std"] = df_g["severe_toxic"]+df_g["toxic"]+df_g["obscene"]+df_g["threat"]+df_g["insult"]+df_g["identity_hate"]

df_g = df_g.fillna(0)


df_g.sort_values(by="total")

jc.query("group_id == 1093")

df_g.query("severe_toxic != 0.0").sort_values(by="total")

jc.query("group_id == 959") 

jc.query("group_id == 83") 

df_g.query("toxic != 0.0").sort_values(by="total")

jc.query("group_id == 1139 and toxic != 0")

jc.to_csv("../processed/jc.csv", index=False)

jc.head()
