==> notebooks/04c-jrsotc-feature-extraction-tf-idf-jc.ipynb <==
import numpy as np

import pandas as pd
pd.set_option('display.max_colwidth', 512)


from sklearn.linear_model import LogisticRegression
from sklearn.feature_extraction.text import TfidfVectorizer
from iterstrat.ml_stratifiers import MultilabelStratifiedKFold
from sklearn.metrics import roc_auc_score
import pickle

### clean data

### build jc train

features = ["toxic","severe_toxic","obscene","threat","insult","identity_hate"]

jc_df = pd.read_csv("../processed/jc.csv")

jc_df.head()

g = jc_df.groupby(["group_id"]).agg({"id":"count","toxic":"std","severe_toxic":"std","obscene":"std","threat":"std","insult":"std","identity_hate":"std"}).fillna(0)
g["std"] = g [features].sum(axis=1)
g.query("std != 0").sort_values(by="id", ascending = False)

jc_df.query("group_id == 1139").sort_values(by="toxic_flag", ascending=False)

jc_df.loc[152180, "group_id"] = jc_df["group_id"].max()
jc_df.loc[34184, "group_id"] = jc_df["group_id"].max()


jc_df.query("group_id == 1139").sort_values(by="toxic_flag", ascending=False)

jc_df.query("group_id == 788").sort_values(by="toxic_flag", ascending=False)

jc_df.query("group_id == 481").sort_values(by="toxic_flag", ascending=False)

g = jc_df.groupby(["group_id"]).agg({"toxic":"max","severe_toxic":"max","obscene":"max","threat":"max","insult":"max","identity_hate":"max"}).reset_index()
print(f"jc_df:{jc_df.shape[0]}")
jc_df = jc_df[["group_id", "id", "text"]].drop_duplicates (subset=["group_id"]).reset_index()
print(f"jc_df:{jc_df.shape[0]}")
jc_df = jc_df.merge(g, on="group_id") 
print(f"jc_df:{jc_df.shape[0]}")
jc_df.head()

g = jc_df.groupby(["group_id"]).agg({"id":"count","toxic":"std","severe_toxic":"std","obscene":"std","threat":"std","insult":"std","identity_hate":"std"}).fillna(0)
g["std"] = g [features].sum(axis=1)
g.query("std != 0").sort_values(by="id", ascending = False)

N_FOLDS = 5
mkf = MultilabelStratifiedKFold(n_splits=N_FOLDS, random_state=123, shuffle=True)
y = jc_df [  ["toxic","severe_toxic","obscene","threat","insult","identity_hate"] ]

jc_df["kfold"] = -1
for f, (t_, v_) in enumerate(mkf.split(X=jc_df, y=y)):
    jc_df.loc[v_, 'kfold'] = f
jc_df.to_csv("../processed/jc.csv", index=False)
jc_df.head()

### Build TF-IDF models 

def subsample_majority_class (df, score_column, perc = 1.5):
    min_len = (df[score_column] > 0).sum()
    df_y0_undersample = df[df[score_column] == 0].sample(n=int(min_len*perc),random_state=201)
    df = pd.concat([df[df[score_column] > 0], df_y0_undersample])
                                                
    return df

def cv_tfidf_predict (model_name, df, score_column):

    pred = np.zeros ((df.shape[0]))

    for fold in range(5):
        vec = TfidfVectorizer(analyzer='char_wb', max_df=0.5, min_df=3, ngram_range=(3, 5)) 

        train = df.query ("kfold != @fold")
        
        valid = df.query ("kfold == @fold")

        X_train = vec.fit_transform(train['text'])
        y_train = train[score_column].values

        X_valid = vec.transform(valid['text'])
        y_valid = valid[score_column].values

        model = Ridge (alpha=1.0)
        model.fit(X_train, y_train)    
        pred_train = model.predict ( X_train )
        pred_valid = model.predict ( X_valid )

        pred [valid.index.values] = pred_valid
        
        
        score_train = mean_squared_error(y_train, pred_train, squared=False)
        score_valid = mean_squared_error (y_valid, pred_valid, squared=False )
        print (f"fold:{fold} rmse_train:{score_train:.5f}, rmse_valid:{score_valid:.5f}")
        pickle.dump(model, open(f"../models/{model_name}_{fold}.pkl", 'wb'))
        pickle.dump(vec, open(f"../models/{model_name}_vec_{fold}.pkl", 'wb'))
        


    y_true = df[score_column]
    score_tot = mean_squared_error ( y_true, pred, squared = False)
    print (f"tot score_tot:{score_tot:.5f}")
    return pred


def cv_tfidf_proba (model_name, df, score_column, subsample_perc = 1.5):

    pred = np.zeros ((df.shape[0]))

    for fold in range(5):
        vec = TfidfVectorizer(analyzer='char_wb', max_df=0.5, min_df=3, ngram_range=(3, 5)) 

        train = df.query ("kfold != @fold")
        
        valid = df.query ("kfold == @fold")

        train = subsample_majority_class (train, score_column,  perc = subsample_perc )

        X_train = vec.fit_transform(train['text'])
        y_train = train[score_column].values

        X_valid = vec.transform(valid['text'])
        y_valid = valid[score_column].values

        model = LogisticRegression ()
        model.fit(X_train, y_train)    
        pred_train = model.predict_proba ( X_train )[:,1]
        pred_valid = model.predict_proba ( X_valid )[:,1]

        pred [valid.index.values] = pred_valid
        
        
        auc_train = roc_auc_score(y_train, pred_train)
        auc_valid = roc_auc_score (y_valid, pred_valid )
        print (f"fold:{fold} auc_train:{auc_train:.5f}, auc_valid:{auc_valid:.5f}")
        pickle.dump(model, open(f"../models/{model_name}_{fold}.pkl", 'wb'))
        pickle.dump(vec, open(f"../models/{model_name}_vec_{fold}.pkl", 'wb'))
        


    y_true = df[score_column]
    auc_tot = roc_auc_score ( y_true, pred)
    print (f"tot auc_tot:{auc_tot:.5f}")
    return pred


def tf_idf_feature (texts, feature_name, path, predict_proba = True,  folds=5):
    pred = np.zeros( (len(texts), ) )
    for fold in range(folds):
        model = pickle.load(open(f"{path}/{feature_name}_{fold}.pkl", 'rb'))
        vec = pickle.load(open(f"{path}/{feature_name}_vec_{fold}.pkl", 'rb'))
        X = vec.transform ( texts )
        if predict_proba:
            pred += model.predict_proba ( X )[:,1]
        else:
            pred += model.predict ( X )
            
    return pred/folds




for feat in features:
    print(feat)
    _ = cv_tfidf_proba(model_name=f"jc_tfidf_{feat}", df=jc_df, score_column=feat, subsample_perc = 1.5 )

%%time

rud = pd.read_csv("../processed/rud_text.csv")

texts = rud["text"].values
for feat in features:
    feat_name = f"jc_tfidf_{feat}"
    print("rud",feat_name)
    rud[feat_name] = tf_idf_feature(texts, feat_name, path="../models", folds=5)

rud.to_csv("../processed/rud_text_jc_tfidf_fe.csv", index=False)


validation_texts = pd.read_csv ("../processed/validation_texts.csv")
texts = validation_texts["text"].values
for feat in features:
    feat_name = f"jc_tfidf_{feat}"
    print("valid",feat_name)
    validation_texts[feat_name] = tf_idf_feature(texts, feat_name, path="../models", folds=5)

validation_texts.to_csv("../processed/valid_text_jc_tfidf_fe.csv", index=False)

validation_texts.head()
